refactor(test): Consolidate scheduler fixtures to conftest.py
Move shared scheduler test fixtures from individual test files to
Tests/conftest.py for reuse:
- temp_schedules_env: Mock schedule storage directories
- sample_schedule_factory: Create test schedules with various triggers
- mock_cron_system: Mock apply_to_system/remove_from_system
- mock_rtc_functions: Mock RTC wakealarm functions

Each test file now only defines its module-specific scheduler_service
fixture that composes these shared fixtures with additional mocking
as needed.
